requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
pillow-avif-plugin>=1.4.0
ultralytics>=8.3.0
opencv-python>=4.8.0
networkx>=3.0
scipy>=1.10.0
scikit-image>=0.21.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
jinja2>=3.1.0
